Revert "[TRTLLM-12500][feat] Add support for Qwen3.5 VL MoE (#14164)"
This reverts commit 96a4a0937e37218a560d6a0e7057dd3ea3ed290c.
