+++ tensorrt_llm/_torch/models/modeling_qwen3_next.py
@@ -973,18 +973,9 @@ class Qwen3NextForCausalLM(SpecDecOneEngineForCausalLM[Qwen3NextModel,
         # is supported for Mamba/SSM-based models
         return {"kv_cache_config": {"enable_block_reuse": False}}
 
-    def load_weights(self,
-                     weights: dict,
-                     weight_mapper: BaseWeightMapper,
-                     params_map: Optional[Dict[str, str]] = None,
-                     allow_partial_loading: bool = False):
+    def load_weights(self, weights: dict, weight_mapper: BaseWeightMapper):
         new_weights = weight_mapper.preprocess_weights(weights)
-        super().load_weights(
-            new_weights,
-            weight_mapper=weight_mapper,
-            params_map=params_map,
-            allow_partial_loading=allow_partial_loading,
-        )
+        super().load_weights(new_weights, weight_mapper)
 
     def post_load_weights(self):
         for idx, layer in enumerate(
