+++ tests/integration/defs/accuracy/references/mmmu.yaml
@@ -62,12 +62,8 @@ Qwen/Qwen3-VL-8B-Instruct:
 mistralai/Mistral-Small-3.1-24B-Instruct-2503:
   - accuracy: 57.0
 Qwen/Qwen3.5-35B-A3B:
-  # The default accuracy for `test_auto_dtype` tests.
-  - accuracy: 59.0
   - dtype: bfloat16
     accuracy: 60.444
-  - quant_algo: FP8_BLOCK_SCALES
-    accuracy: 58.889
 # Kimi K2.5 multimodal (MoonViT + DeepSeek-V3 MoE backbone, ~1T params).
 # Values below are measured with NVFP4 checkpoint (thinking mode enabled).
 moonshotai/Kimi-K2.5:
