+++ tests/unittest/_torch/modeling/test_modeling_multimodal.py
@@ -18,12 +18,6 @@ from tensorrt_llm._torch.attention_backend.utils import get_attention_backend
 from tensorrt_llm._torch.metadata import KVCacheParams
 from tensorrt_llm._torch.model_config import ModelConfig
 from tensorrt_llm._torch.models.modeling_multimodal_utils import bypass_processor_output_validation
-from tensorrt_llm._torch.pyexecutor.config_utils import (
-    extract_mamba_kv_cache_params,
-    is_nemotron_hybrid,
-    is_qwen3_hybrid,
-)
-from tensorrt_llm._torch.pyexecutor.mamba_cache_manager import CppMambaHybridCacheManager
 from tensorrt_llm._torch.pyexecutor.resource_manager import KVCacheManager
 from tensorrt_llm._utils import str_dtype_to_torch
 from tensorrt_llm.bindings.executor import KvCacheConfig
@@ -34,7 +28,6 @@ from tensorrt_llm.inputs import (
     prompt_inputs,
 )
 from tensorrt_llm.inputs.multimodal import MultimodalParams, MultimodalRuntimeData
-from tensorrt_llm.llmapi.llm_args import KvCacheConfig as PyKvCacheConfig
 from tensorrt_llm.mapping import Mapping
 
 
@@ -525,13 +518,6 @@ class TestModelingMultimodal(unittest.TestCase, ABC):
         Note:
             This method uses get_kv_cache_config() to obtain configuration.
             Override get_kv_cache_config() to customize cache settings.
-
-            For hybrid linear-attention models (Qwen3Next, Qwen3.5,
-            Nemotron-Hybrid) this dispatches to
-            `get_hybrid_kv_cache_manager` so the linear-attention layers
-            get a `CppMambaHybridCacheManager` for SSM/conv state.
-            Mirrors the production dispatch in
-            `_util.py:_create_kv_cache_manager`.
         """
         # Get cache configuration from the configurable method
         cache_config = self.get_kv_cache_config(scenario)
@@ -541,113 +527,16 @@ class TestModelingMultimodal(unittest.TestCase, ABC):
 
         num_blocks = (max_seq_len + tokens_per_block - 1) // tokens_per_block
 
-        config = self.model_config.pretrained_config
-        text_config = getattr(config, "text_config", config)
-
-        if is_qwen3_hybrid(text_config) or is_nemotron_hybrid(text_config):
-            self.kv_cache_manager = self.get_hybrid_kv_cache_manager(
-                text_config=text_config,
-                tokens_per_block=tokens_per_block,
-                max_seq_len=max_seq_len,
-                batch_size=batch_size,
-                num_blocks=num_blocks,
-            )
-        else:
-            self.kv_cache_manager = self.get_kv_cache_manager(
-                dtype=self.model_config.pretrained_config.torch_dtype,
-                config=self.model_config.pretrained_config,
-                tokens_per_block=tokens_per_block,
-                max_seq_len=max_seq_len,
-                batch_size=batch_size,
-                num_blocks=num_blocks,
-            )
-
-        self.kv_cache_manager.add_dummy_requests(
-            request_ids=[1],
-            token_nums=[max_seq_len],
-            **self._dummy_request_kwargs(scenario),
-        )
-
-    def _dummy_request_kwargs(self, scenario: MultimodalScenario) -> Dict:
-        """Optional override hook for extra kwargs to `add_dummy_requests`.
-
-        Subclasses for mRoPE-using models (Qwen2.5-VL, Qwen3-VL, Qwen3.5-VL,
-        …) should return `{"use_mrope": True}` here so the cache manager
-        allocates the mRoPE position-id buffer at dummy-request time.
-        Defaults to an empty dict, preserving existing behavior for tests
-        that don't care.
-        """
-        return {}
-
-    def get_hybrid_kv_cache_manager(
-        self,
-        text_config: PretrainedConfig,
-        tokens_per_block: int,
-        max_seq_len: int,
-        batch_size: int,
-        num_blocks: int,
-    ):
-        """Build a `CppMambaHybridCacheManager` for hybrid linear-attention
-        models (Qwen3Next, Qwen3.5, Nemotron-Hybrid).
-
-        Mirrors the production construction in
-        `_util.py:_create_kv_cache_manager` for `is_qwen3_hybrid` /
-        `is_nemotron_hybrid` configs: pulls the state-shape / dtype /
-        layer-mask parameters from `extract_mamba_kv_cache_params` and
-        threads them through the constructor. Tests that need a different
-        concrete manager (e.g. `MixedMambaHybridCacheManager` for
-        disagg-style coverage) should override this method.
-        """
-        dtype_map = {
-            torch.half: tensorrt_llm.bindings.DataType.HALF,
-            torch.float16: tensorrt_llm.bindings.DataType.HALF,
-            torch.bfloat16: tensorrt_llm.bindings.DataType.BF16,
-        }
-
-        mamba_params = extract_mamba_kv_cache_params(text_config)
-        if mamba_params.dtype not in dtype_map:
-            raise ValueError(
-                f"Unsupported dtype for hybrid cache manager: "
-                f"{mamba_params.dtype}. Supported: {list(dtype_map.keys())}"
-            )
-        kv_cache_dtype = dtype_map[mamba_params.dtype]
-
-        head_dim = getattr(text_config, "head_dim", None)
-        if not isinstance(head_dim, int):
-            head_dim = text_config.hidden_size // text_config.num_attention_heads
-
-        # CppMambaHybridCacheManager reads Pydantic-only fields
-        # (mamba_state_cache_interval, enable_block_reuse) so we have to
-        # construct the llmapi.llm_args.KvCacheConfig here, not the C++
-        # bindings KvCacheConfig that the standard KVCacheManager path uses.
-        kv_cache_config = PyKvCacheConfig(max_tokens=num_blocks * tokens_per_block)
-        mapping = Mapping(world_size=1, tp_size=1, rank=0)
-
-        return CppMambaHybridCacheManager(
-            # mamba cache parameters (positional)
-            mamba_params.state_size,
-            mamba_params.conv_kernel,
-            mamba_params.num_heads,
-            mamba_params.n_groups,
-            mamba_params.head_dim,
-            mamba_params.num_mamba_layers,
-            mamba_params.mamba_layer_mask,
-            mamba_params.dtype,
-            mamba_params.mamba_ssm_cache_dtype,
-            # kv cache parameters (positional)
-            kv_cache_config,
-            tensorrt_llm.bindings.internal.batch_manager.CacheType.SELF,
-            # kw-only
-            num_layers=mamba_params.num_full_attention_layers,
-            layer_mask=mamba_params.full_attention_layer_mask,
-            num_kv_heads=text_config.num_key_value_heads,
-            head_dim=head_dim,
+        self.kv_cache_manager = self.get_kv_cache_manager(
+            dtype=self.model_config.pretrained_config.torch_dtype,
+            config=self.model_config.pretrained_config,
             tokens_per_block=tokens_per_block,
             max_seq_len=max_seq_len,
-            max_batch_size=batch_size,
-            mapping=mapping,
-            dtype=kv_cache_dtype,
+            batch_size=batch_size,
+            num_blocks=num_blocks,
         )
+
+        self.kv_cache_manager.add_dummy_requests(request_ids=[1], token_nums=[max_seq_len])
 
     def get_max_num_tokens(self, scenario: MultimodalScenario) -> int:
         """Get maximum number of tokens for attention metadata."""
@@ -805,14 +694,6 @@ class TestModelingMultimodal(unittest.TestCase, ABC):
 
         # TODO: Add multi-GPU support
         self.device = torch.device("cuda:0")
-
-        # Pre-initialize fields that tearDown / setup_scenario expect to
-        # exist. Without this, a test method that doesn't run
-        # setup_scenario (e.g. a setUp-only smoke test) leaves
-        # self.kv_cache_manager unset and tearDown errors with
-        # AttributeError on the ``is not None`` check.
-        self.kv_cache_manager = None
-        self.attn_metadata = None
 
         self.hf_config = self.create_hf_config()
         if self.skip_hf_inference:
